+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.3.3]
+
+Support for Satodime v0.2+:
+* Add SatodimeStatus isFixedCvc & isCoa fields
+* Update repo for package secp256k1
+* Update repo for package CryptoSwift
+
 ## [0.3.2]
 
 Add ECDSA, Schnorr & Musig2 signatures support for Satochip: